+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# 0.1.3
+Fix critical bug in method `Action.empty`  
+small refactoring action.js, selector.js files  
+Write new tests   
+Update documentation  
+
 ## 0.1.2
 Write new tests (increase coverage)  
 Add selector `getMergedActionsData`, for details please see [source](https://github.com/edtoken/redux-tide/blob/master/src/selector.js#L116) and [example](https://edtoken.github.io/redux-tide/?ex=merged-actions-data) and [example-source](https://github.com/edtoken/redux-tide/blob/master/website/src/merged-actions-data/index.js#L50)